mixxx: Add option for library location/indexing.
Sets the default library path to $HOME/Music of the DJ user, because on
the actual production system, the directory contains a symlink pointing
to the storage while containing local files as well.

Apart from the library path the options default to how it is on the
production system, so the main reason for introducing them is VM
testing.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>