Add simple test playing sine waves on both decks.
Nothing too spectacular here, except the method we're using for testing,
as it's quite tricky to check whether Mixxx has been started up
correctly _and_ also is ready to handle keyboard input. Which it is not
right when the window appears, but a few ms after the library scan has
completed.

Of course, this test would be way better without the sleep xdotool
commands, but I haven't found a better implementation yet.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>